requirements.txt
streamlit
tensorflow
scikit-learn
joblib
pillow